add federated adaptive methods in separate files
