+++ src/appfl/algorithm/fedadapt_avg.py
@@ -1,0 +1,6 @@
+from .fedadapt import FedAdaptServer
+
+class FedAvgServer(FedAdaptServer):
+    def compute_step(self):        
+        for name, _ in self.model.named_parameters():                        
+            self.step[name] = self.pseudo_grad[name]
