+++ src/appfl/algorithm/fedadapt_avgm.py
@@ -1,0 +1,7 @@
+from .fedadapt import FedAdaptServer
+
+class FedAvgMServer(FedAdaptServer):
+    def compute_step(self):        
+        super(FedAvgMServer, self).update_m_vector()
+        for name, _ in self.model.named_parameters():                        
+            self.step[name] = self.m_vector[name]
